Speed up check for positive instance before evaluation
For some reason only checking until the first positive instance is found
changes the behaviour of GradientSearch although non of the previously
collected samples is used.

I could not find a reason for this bevahiour and since it still works
correctly afterwards, I decided to commit this. It's alot faster.